+++ app/helpers/requerimientos_helper.rb
@@ -3,4 +3,5 @@ module RequerimientosHelper
     aprobado = requerimiento.presupuestos.exists? ? requerimiento.presupuestos.aprobado : nil
     aprobado.con_iva? ? ' (*)' : ' ()' if aprobado      
   end  
+
 end