+++ CHANGELOG.md
@@ -4,3 +4,17 @@
 * Se agrega el orden por "Estado" en la lista de requerimientos. 
 * Se agrega el filtro de razón social del proveedor en la pantalla de requerimientos.  Se filtrarán solo los requerimientos que ya hayan sido apro    bados por compras (o sea, que poseen un presupuesto asignado). 
 * Se agrega el campo "Consumo" en los requerimientos.  El consumo puede ser una opción dentro de "semanal, quincenal, bimestral, trimestral, semestral y anual" 
+
+## Version 0.1.2
+
+* El listado de requerimientos puede ser exportado a excel juntos con sus presupuestos aprobados (Issue #142)
+
+## Version 0.1.5
+
+* Errores #146: Ocurre un error al intentar aprobar algunos requerimientos por compras
+* Funcionalidad #145: Consumo, se agrega la opción 'eventual'
+
+## Version 0.1.6
+
+* Error #150: No se muestran todos los usuario
+* Funcionalidad #144: Aprobado de requerimiento en simultaneo